requirements.txt
annotated-types
anyio
certifi
click
dnspython
fastapi
greenlet
h11
httpcore
httpx
idna
iniconfig
mongoengine
motor
packaging
pluggy
pydantic
pydantic_core
pymongo
pytest
python-dotenv
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
bcrypt
PyJWT
pytest-asyncio
freezegun
pydantic[email]
beanie
